feat: load pre-context ahead of user message
Prepend pre-context message ahead of user message and benefit from
prompt caching. Consider this poor man's RAG.
